Move class name generation to fuselage
